- Refactoring ViewActions and api - Error handling for login errors
